arl/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
black==23.3.0
blosc2==2.5.1
boto==2.49.0
boto3==1.28.25
botocore==1.31.25
Brotli==1.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.2.7
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==42.0.4
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.10
django-appconf==1.0.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-countries==7.5.1
django-crispy-forms==2.0
django-cryptography-django5==2.2
django-htmx==1.15.0
django-import-export==3.2.0
django-intl-tel-input==0.3.1
django-js-asset==2.2.0
django-phonenumber-field==7.1.0
django-ratelimit==4.1.0
django-sendgrid-v5==1.2.2
django-taggit==4.0.0
django-timezone-field==5.1
django-waffle==4.1.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.14.0
docusign-esign==3.23.0
dropbox==11.36.2
et-xmlfile==1.1.0
exceptiongroup==1.1.2
factory-boy==3.2.1
Faker==18.11.2
flake8==6.0.0
flower==2.0.0
fonttools==4.46.0
freetype-py==2.3.0
frozenlist==1.3.3
h5py==3.11.0
html5lib==1.1
humanize==4.7.0
idna==3.4
importlib_metadata==8.5.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.3.1
lml==0.1.0
lxml==4.9.3
MarkupPy==1.14
mccabe==0.7.0
msgpack==1.1.0
multidict==6.0.4
mypy-extensions==1.0.0
ndindex==1.9
nose==1.3.7
numexpr==2.10.1
numpy==1.26.0
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==24.2
pandas==2.2.2
pathspec==0.11.1
pdfkit==1.0.0
phonenumbers==8.13.15
pillow==10.2.0
pipdeptree==2.24.0
platformdirs==3.8.0
pluggy==1.2.0
ply==3.11
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
pycairo==1.25.1
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.8.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyflakes==3.0.1
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.7.0
pypdf==3.15.2
pyphen==0.14.0
pypng==0.20220715.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==4.0.7
requests==2.31.0
rlPyCairo==0.3.0
s3transfer==0.6.1
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stats-can==2.5.3
stone==3.3.1
svglib==1.5.1
tables==3.9.2
tablib==3.5.0
texttable==1.7.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.66.5
twilio==8.4.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
xhtml2pdf==0.2.13
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
zipp==3.20.2
zopfli==0.2.3
